Create Containers directory.
